Reorganize glsl functions
